requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
MarkupSafe==3.0.2
pyasn1==0.6.1
python-dotenv==1.1.1
Werkzeug==3.1.3
